requirements.txt
requests
selenium
Pillow
beautifulsoup4
colorama
